+++ forui/test/src/widgets/time_picker/time_picker_test.dart
@@ -100,7 +100,6 @@ void main() {
         ),
       );
 
-      expect(initial.disposed, false);
       expect(current.value, const FTime(10, 30));
       expect(current.picker?.value, [10, 30, 0]);
     });