+++ forui/CHANGELOG.md
@@ -163,9 +163,9 @@ are not available for most of these changes due to the tool's limitations.
   Wrap existing values with `.all(...)`, e.g. `.all(IconThemeData(...))`.
 * **Breaking** Rename `FSelectStyle.selectFieldStyle` to `FSelectStyle.fieldStyle`.
 * **Breaking** Rename `FSelectSearchStyle.textFieldStyle` to `FSelectSearchStyle.fieldStyle`.
-
 * **Breaking** Remove `FSelectStyle.iconStyle`. Use `FSelectStyle.fieldStyle.iconStyle` instead.
 * **Breaking** Remove `FSelectSearchStyle.iconStyle`. Use `FSelectSearchStyle.fieldStyle.iconStyle` instead.
+* Restore `FSelect.contentScrollHandle`.
 
 * Fix `FMultiSelect` still allowing tags to be removed when disabled.
 
@@ -261,6 +261,7 @@ are not available for most of these changes due to the tool's limitations.
 * **Breaking** Replace all instances of `FWidgetStateMap` with `FVariants`.
 * **Breaking** Remove `call` method from all styles.
 * **Breaking** Remove `FWidgetStatesDelta`.
+* **Breaking** Remove `FValueNotifier` - use `ValueNotifier` instead.
 
 * Fix CLI incorrectly parsing `FStyle`.